Moved project dependencies update from "afterEvaluate" to direct evaluation in order to support gradle multi-projects. Cf. https://discuss.gradle.org/t/when-can-a-plugin-change-a-projects-dependencies/17505 and https://www.oreilly.com/library/view/gradle-beyond-the/9781449373801/ch03.html
